requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
boto3==1.35.24
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1